secured admin functions

+++ src/main/java/com/servlet/admin/DeleteServlet.java
@@ -4,6 +4,8 @@ import com.bean.Timetable;
 import com.service.TimeTableService;
 import com.service.impl.TimeTableServiceImpl;
 import com.util.BeanUtils;
+import com.util.SecuredUtil;
+
 import jakarta.servlet.http.HttpServlet;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
@@ -13,23 +15,27 @@ import java.io.PrintWriter;
 import java.util.logging.Logger;
 
 public class DeleteServlet extends HttpServlet {
-    private static final long serialVersionUID = 1L;
+	private static final long serialVersionUID = 1L;
 
-    private final Logger log = Logger.getLogger(DeleteServlet.class.getName());
+	private final Logger log = Logger.getLogger(DeleteServlet.class.getName());
 
-    private final TimeTableService timeTableService = new TimeTableServiceImpl();
+	private final TimeTableService timeTableService = new TimeTableServiceImpl();
 
-    public void doPost(HttpServletRequest req, HttpServletResponse res) throws IOException {
-        log.info("Request to delete timetable data");
-        Timetable bean = new Timetable();
-        BeanUtils.populate(bean, req);
-        String collegeName = req.getParameter("collegeName");
-        int status = timeTableService.delete(bean, collegeName);
-        if (status == 404) {
-            PrintWriter out = res.getWriter();
-            out.print("Invalid record!");
-        }
-        res.setStatus(status);
-    }
+	public void doPost(HttpServletRequest req, HttpServletResponse res) throws IOException {
+		log.info("Request to delete timetable data");
+		if (SecuredUtil.allow(req.getSession(false))) {
+			Timetable bean = new Timetable();
+			BeanUtils.populate(bean, req);
+			String collegeName = req.getParameter("collegeName");
+			int status = timeTableService.delete(bean, collegeName);
+			if (status == 404) {
+				PrintWriter out = res.getWriter();
+				out.print("Invalid record!");
+			}
+			res.setStatus(status);
+		} else {
+			res.setStatus(HttpServletResponse.SC_FORBIDDEN);
+		}
+	}
 
 }
